working to select detector, update plot
if you select the text of the detectors in the left hand axis, you can
move up and down the length of the freeway.

perhaps I should implement a log scale, because sometimes the distances
get huge

lots of tedious trial and error to get the transitions to work properly,
but it seems to be correct now.  When selecting a new detector, the old
plot is removed, new plot drawn, without the hacky "remove()" commands I
have in the grid_display client codes.